requirements.txt
# -------------------------
# Core Framework
# -------------------------
Flask==3.0.3
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.6.0

# -------------------------
# Database & ORM
# -------------------------
SQLAlchemy==2.0.32
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9   # PostgreSQL driver
alembic==1.13.2          # database migrations

# -------------------------
# Validation & Security
# -------------------------
passlib==1.7.4           # password hashing
email-validator==2.1.1   # validate emails
python-dotenv==1.0.1     # load .env files
cryptography==43.0.1     # secure tokens, encryption

# -------------------------
# Background Jobs & Scheduling
# -------------------------
celery==5.4.0            # async tasks (optional but recommended)
redis==5.0.4             # Celery broker

# External API Requests
requests>=2.8.1,<3.0
aiohttp==3.9.5
python-mpesa==0.1.10
PyJWT==2.8.0
          # generate tokens for secure API calls

# -------------------------
# Utils & Helpers
# -------------------------
numpy>=1.24             # useful for scoring, math
pandas>=2.0           # helpful for reporting
python-dateutil==2.9.0   # date parsing
pytz==2024.1             # timezone conversion
gunicorn==22.0.0         # production WSGI server

# -------------------------
# Logging & Monitoring
# -------------------------
loguru==0.7.2

# -------------------------
# Testing
# -------------------------
pytest==8.3.2
pytest-cov>=5.0

# -------------------------
# Security Hardening
# -------------------------
Werkzeug==3.0.2
charset-normalizer==3.3.2
itsdangerous==2.2.0

# -------------------------
# Serialization
# -------------------------
marshmallow==3.21.3      # schema validation
marshmallow-sqlalchemy==1.1.0

# -------------------------
# Caching & Performance
# -------------------------
Flask-Caching==2.2.0

# -------------------------
# Production Deployment
# -------------------------
uvicorn==0.30.1          # optional ASGI server
